pyproject.toml
[project]
name = "jazz-repertoire"
version = "1.0.0"
description = "A jazz repertoire management app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-allauth>=65.7.0",
    "gunicorn>=23.0.0",
    "psycopg2>=2.9.10",
    "python-decouple>=3.8",
    "sendgrid>=6.11.0",
    "sentry-sdk>=2.27.0",
    "sphinx>=8.2.3",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
    "ipdb>=0.13.13",
    "playwright>=1.52.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-playwright>=0.7.0",
    "tox>=4.26.0",
]
